+++ Changelog.md
@@ -1,3 +1,8 @@
 0.7:
 - minor bugfixes
 - HAOS can now be installed in a VM
+
+0.8:
+- Optimized for Electric Eel (24.10)
+    - No docker install option anymore, since 24.10 comes with docker out of the box
+    - Adapted install mode to modify new installer script